# MRG, ENH: Speed up src filtering and vol label src creation