Closes #5123.

This was a fun one! @agramfort your original example with `pos=2.` is a bit crazy, it creates 342079. So I tested it with `pos=5.`, which took 36 sec on master:
<details>

```
import os.path as op
import time

from mne.datasets import sample
from mne import setup_volume_source_space

data_path = sample.data_path()
subject = 'sample'
data_dir = op.join(data_path, 'MEG', subject)
subjects_dir = op.join(data_path, 'subjects')
bem_dir = op.join(subjects_dir, subject, 'bem')
fname_aseg = op.join(subjects_dir, subject, 'mri', 'aseg.mgz')
fname_model = op.join(bem_dir, '%s-5120-bem.fif' % subject)
labels_vol = ['Left-Hippocampus',
              'Right-Hippocampus']


def test():
    t0 = time.time()
    # create a mixed src space
    hippocampus = setup_volume_source_space(
        subject,
        mri=fname_aseg,
        pos=5.0,
        bem=fname_model,
        volume_label=labels_vol,
        subjects_dir=subjects_dir,
        verbose=True)
    print(time.time() - t0)


if __name__ == '__main__':
    test()
```
</details>

I cut this in half (18 sec) by:

1. Only checking inside-ness for points that are not clearly inside (fit inside a sphere nested within the mesh)
2. Using `scipy.spatial.Delaunay` (or equivalently `ConvexHull`) to eliminate points that are not within the convex hull of the surface before using the solid angles check.
3. Only checking `min_dist` for points that are inside at all
4. Only checking `min_dist` for those that can possibly exceed it (again using the sphere inside the mesh)
5. Checking `min_dist` using `cKDTree` instead of `BallTree`

This will speed up any `_filter_source_spaces` call (including those in our `fwd` code).

Then I realized that the order of operations for volume labels was:

1. Create the volume source space for the entire `inner_skull` region (including point filtering!)
2. Cull points based on the `aseg` proximity

So if you had two labels, `_filter_source_spaces` was not only called twice, but it operated on the 342079 points instead of the 450 and 516 (for `pos=2.`) that were in the L/R Hip (!). Switching the order of these two made the `pos=5.` code run in 3.3 sec, and even the much tighter `pos=2.0` spacing code run in 4.6 sec instead of taking an eternity.

I also refactored the logging a bit to make it much more readable:

<details>

```
larsoner@bunk:~/Desktop$ python -ui vol.py 
BEM              : /home/larsoner/mne_data/MNE-sample-data/subjects/sample/bem/sample-5120-bem.fif
grid                  : 2.0 mm
mindist               : 5.0 mm
MRI volume            : /home/larsoner/mne_data/MNE-sample-data/subjects/sample/mri/aseg.mgz

Loaded inner skull from /home/larsoner/mne_data/MNE-sample-data/subjects/sample/bem/sample-5120-bem.fif (2562 nodes)
Surface CM = (   0.7  -10.0   44.3) mm
Surface fits inside a sphere with radius   91.8 mm
Surface extent:
    x =  -66.7 ...   68.8 mm
    y =  -88.0 ...   79.0 mm
    z =  -44.5 ...  105.8 mm
Grid extent:
    x =  -68.0 ...   70.0 mm
    y =  -90.0 ...   80.0 mm
    z =  -46.0 ...  106.0 mm
463540 sources before omitting any.
342079 sources after omitting infeasible sources not within 0.0 - 91.8 mm.

Selecting voxels from Left-Hippocampus
450 sources remaining after excluding sources too far from VOI voxels
Source spaces are in MRI coordinates.
Checking that the sources are inside the surface and at least    5.0 mm away (will take a few...)
    Skipping interior check for 395 sources that fit inside a sphere of radius   43.6 mm
    Skipping solid angle check for 0 points using Qhull
450 sources remaining after excluding the sources outside the surface and less than    5.0 mm inside.
Adjusting the neighborhood info.

Selecting voxels from Right-Hippocampus
516 sources remaining after excluding sources too far from VOI voxels
Source spaces are in MRI coordinates.
Checking that the sources are inside the surface and at least    5.0 mm away (will take a few...)
    Skipping interior check for 431 sources that fit inside a sphere of radius   43.6 mm
    Skipping solid angle check for 0 points using Qhull
516 sources remaining after excluding the sources outside the surface and less than    5.0 mm inside.
Adjusting the neighborhood info.

Reading /home/larsoner/mne_data/MNE-sample-data/subjects/sample/mri/aseg.mgz...
Source space : MRI voxel -> MRI (surface RAS)
     0.002000  0.000000  0.000000     -68.00 mm
     0.000000  0.002000  0.000000     -90.00 mm
     0.000000  0.000000  0.002000     -46.00 mm
     0.000000  0.000000  0.000000       1.00
MRI volume : MRI voxel -> MRI (surface RAS)
    -0.001000  0.000000  0.000000     128.00 mm
     0.000000  0.000000  0.001000    -128.00 mm
     0.000000 -0.001000  0.000000     128.00 mm
     0.000000  0.000000  0.000000       1.00
MRI volume : MRI (surface RAS) -> RAS (non-zero origin)
     1.000000 -0.000000 -0.000000      -5.27 mm
    -0.000000  1.000000 -0.000000       9.04 mm
    -0.000000  0.000000  1.000000     -27.29 mm
     0.000000  0.000000  0.000000       1.00
Setting up interpolation for Left-Hippocampus...
 54848/16777216 nonzero values [done]
Setting up interpolation for Right-Hippocampus...
 62720/16777216 nonzero values [done]
```
</details>