@agramfort 2019-08-23
@larsoner you need to fix conflicts