@agramfort 2019-08-26
thx @larsoner !